requirements.txt
Django==4.0.1
psycopg2-binary==2.9.3
djangorestframework==3.13.1
django-filter==21.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
pytest==7.4.4